Fix: Moved the types file to lib/types
